Restructure into a package rather than 3 separate-looking modules
